requirements.txt
fastapi
uvicorn[standard]
python-multipart
gradio
torch
diffusers
transformers
accelerate
xformers
Pillow
numpy